refactor(indexing): rename interfaces.ts -> types.ts + import Modal from Opal
- Rename web/src/lib/indexing/interfaces.ts -> types.ts (review nit from @raunakab)
  and re-point all 8 importers to @/lib/indexing/types.
- ReindexErrorsModal: import { Modal } from "@opal/components" — the old
  @/refresh-components/Modal was migrated into Opal on main.
